Update to using Cython 0.29.26
Currently attempting to use this package with python 3.10 results in
an error on import along the lines of:

~~~
>>> import heatshrink2
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File ".../python3.10/site-packages/heatshrink2/__init__.py", line 5, in <module>
    from . import core
ImportError: .../python3.10/site-packages/heatshrink2/core.cpython-310-x86_64-linux-gnu.so: undefined symbol: _PyGen_Send
>>>

This is because of some API changes in Python 3.10 that the version of
cython that this package was built with does not support.

Cython 0.29.22 incorporated some fixes to support python 3.10.  The
one that seems particularly relevant to this package is this one:
https://github.com/cython/cython/pull/3921

This patch bumps the cython requirement to the latest version, and
regenerates the cython C files.