Upgrade to using Cython 0.29.33
Attempting to install this package with python 3.11 currently fails
when pip attempts to build the wheel from the sdist. This appears
to be due to changes introduced in CPython 3.11a4 which removed
exc_type and exc_traceback fields in the _PyErr_StackItem structure.
See: https://github.com/cython/cython/issues/4500

This patch bumps the cython requirement to the latest version, and
regenerates the cython C files.